+++ cmd/abci-cli/abci-cli.go
@@ -54,8 +54,8 @@ var (
 
 var RootCmd = &cobra.Command{
 	Use:   "abci-cli",
-	Short: "",
-	Long:  "",
+	Short: "the ABCI CLI tool wraps an ABCI client",
+	Long:  "the ABCI CLI tool wraps an ABCI client and is used for testing ABCI servers",
 	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
 
 		switch cmd.Use {
@@ -92,6 +92,7 @@ type response struct {
 	// generic abci response
 	Data []byte
 	Code uint32
+	Info string
 	Log  string
 
 	Query *queryResponse
@@ -111,26 +112,26 @@ func Execute() error {
 }
 
 func addGlobalFlags() {
-	RootCmd.PersistentFlags().StringVarP(&flagAddress, "address", "", "tcp://0.0.0.0:46658", "Address of application socket")
-	RootCmd.PersistentFlags().StringVarP(&flagAbci, "abci", "", "socket", "Either socket or grpc")
-	RootCmd.PersistentFlags().BoolVarP(&flagVerbose, "verbose", "v", false, "Print the command and results as if it were a console session")
-	RootCmd.PersistentFlags().StringVarP(&flagLogLevel, "log_level", "", "debug", "Set the logger level")
+	RootCmd.PersistentFlags().StringVarP(&flagAddress, "address", "", "tcp://0.0.0.0:46658", "address of application socket")
+	RootCmd.PersistentFlags().StringVarP(&flagAbci, "abci", "", "socket", "either socket or grpc")
+	RootCmd.PersistentFlags().BoolVarP(&flagVerbose, "verbose", "v", false, "print the command and results as if it were a console session")
+	RootCmd.PersistentFlags().StringVarP(&flagLogLevel, "log_level", "", "debug", "set the logger level")
 }
 
 func addQueryFlags() {
-	queryCmd.PersistentFlags().StringVarP(&flagPath, "path", "", "/store", "Path to prefix query with")
-	queryCmd.PersistentFlags().IntVarP(&flagHeight, "height", "", 0, "Height to query the blockchain at")
-	queryCmd.PersistentFlags().BoolVarP(&flagProve, "prove", "", false, "Whether or not to return a merkle proof of the query result")
+	queryCmd.PersistentFlags().StringVarP(&flagPath, "path", "", "/store", "path to prefix query with")
+	queryCmd.PersistentFlags().IntVarP(&flagHeight, "height", "", 0, "height to query the blockchain at")
+	queryCmd.PersistentFlags().BoolVarP(&flagProve, "prove", "", false, "whether or not to return a merkle proof of the query result")
 }
 
 func addCounterFlags() {
-	counterCmd.PersistentFlags().StringVarP(&flagAddrC, "addr", "", "tcp://0.0.0.0:46658", "Listen address")
-	counterCmd.PersistentFlags().BoolVarP(&flagSerial, "serial", "", false, "Enforce incrementing (serial) transactions")
+	counterCmd.PersistentFlags().StringVarP(&flagAddrC, "addr", "", "tcp://0.0.0.0:46658", "listen address")
+	counterCmd.PersistentFlags().BoolVarP(&flagSerial, "serial", "", false, "enforce incrementing (serial) transactions")
 }
 
 func addDummyFlags() {
-	dummyCmd.PersistentFlags().StringVarP(&flagAddrD, "addr", "", "tcp://0.0.0.0:46658", "Listen address")
-	dummyCmd.PersistentFlags().StringVarP(&flagPersist, "persist", "", "", "Directory to use for a database")
+	dummyCmd.PersistentFlags().StringVarP(&flagAddrD, "addr", "", "tcp://0.0.0.0:46658", "listen address")
+	dummyCmd.PersistentFlags().StringVarP(&flagPersist, "persist", "", "", "directory to use for a database")
 }
 func addCommands() {
 	RootCmd.AddCommand(batchCmd)
@@ -155,18 +156,39 @@ func addCommands() {
 
 var batchCmd = &cobra.Command{
 	Use:   "batch",
-	Short: "Run a batch of abci commands against an application",
-	Long:  "",
-	Args:  cobra.ExactArgs(0),
+	Short: "run a batch of abci commands against an application",
+	Long: `run a batch of abci commands against an application
+
+This command is run by piping in a file containing a series of commands
+you'd like to run:
+
+    abci-cli batch < example.file
+
+where example.file looks something like:
+
+    set_option serial on
+    check_tx 0x00
+    check_tx 0xff
+    deliver_tx 0x00
+    check_tx 0x00
+    deliver_tx 0x01
+    deliver_tx 0x04
+    info
+`,
+	Args: cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdBatch(cmd, args)
 	},
 }
 
 var consoleCmd = &cobra.Command{
-	Use:       "console",
-	Short:     "Start an interactive abci console for multiple commands",
-	Long:      "",
+	Use:   "console",
+	Short: "start an interactive ABCI console for multiple commands",
+	Long: `start an interactive ABCI console for multiple commands
+
+This command opens an interactive console for running any of the other commands
+without opening a new connection each time
+`,
 	Args:      cobra.ExactArgs(0),
 	ValidArgs: []string{"echo", "info", "set_option", "deliver_tx", "check_tx", "commit", "query"},
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -176,8 +198,8 @@ var consoleCmd = &cobra.Command{
 
 var echoCmd = &cobra.Command{
 	Use:   "echo",
-	Short: "Have the application echo a message",
-	Long:  "",
+	Short: "have the application echo a message",
+	Long:  "have the application echo a message",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdEcho(cmd, args)
@@ -185,8 +207,8 @@ var echoCmd = &cobra.Command{
 }
 var infoCmd = &cobra.Command{
 	Use:   "info",
-	Short: "Get some info about the application",
-	Long:  "",
+	Short: "get some info about the application",
+	Long:  "get some info about the application",
 	Args:  cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdInfo(cmd, args)
@@ -194,8 +216,8 @@ var infoCmd = &cobra.Command{
 }
 var setOptionCmd = &cobra.Command{
 	Use:   "set_option",
-	Short: "Set an option on the application",
-	Long:  "",
+	Short: "set an option on the application",
+	Long:  "set an option on the application",
 	Args:  cobra.ExactArgs(2),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdSetOption(cmd, args)
@@ -204,8 +226,8 @@ var setOptionCmd = &cobra.Command{
 
 var deliverTxCmd = &cobra.Command{
 	Use:   "deliver_tx",
-	Short: "Deliver a new transaction to the application",
-	Long:  "",
+	Short: "deliver a new transaction to the application",
+	Long:  "deliver a new transaction to the application",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdDeliverTx(cmd, args)
@@ -214,8 +236,8 @@ var deliverTxCmd = &cobra.Command{
 
 var checkTxCmd = &cobra.Command{
 	Use:   "check_tx",
-	Short: "Validate a transaction",
-	Long:  "",
+	Short: "validate a transaction",
+	Long:  "validate a transaction",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdCheckTx(cmd, args)
@@ -224,8 +246,8 @@ var checkTxCmd = &cobra.Command{
 
 var commitCmd = &cobra.Command{
 	Use:   "commit",
-	Short: "Commit the application state and return the Merkle root hash",
-	Long:  "",
+	Short: "commit the application state and return the Merkle root hash",
+	Long:  "commit the application state and return the Merkle root hash",
 	Args:  cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdCommit(cmd, args)
@@ -234,8 +256,8 @@ var commitCmd = &cobra.Command{
 
 var versionCmd = &cobra.Command{
 	Use:   "version",
-	Short: "Print abci console version",
-	Long:  "",
+	Short: "print ABCI console version",
+	Long:  "print ABCI console version",
 	Args:  cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		fmt.Println(version.Version)
@@ -245,8 +267,8 @@ var versionCmd = &cobra.Command{
 
 var queryCmd = &cobra.Command{
 	Use:   "query",
-	Short: "Query the application state",
-	Long:  "",
+	Short: "query the application state",
+	Long:  "query the application state",
 	Args:  cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdQuery(cmd, args)
@@ -256,7 +278,7 @@ var queryCmd = &cobra.Command{
 var counterCmd = &cobra.Command{
 	Use:   "counter",
 	Short: "ABCI demo example",
-	Long:  "",
+	Long:  "ABCI demo example",
 	Args:  cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdCounter(cmd, args)
@@ -266,7 +288,7 @@ var counterCmd = &cobra.Command{
 var dummyCmd = &cobra.Command{
 	Use:   "dummy",
 	Short: "ABCI demo example",
-	Long:  "",
+	Long:  "ABCI demo example",
 	Args:  cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdDummy(cmd, args)
@@ -275,8 +297,8 @@ var dummyCmd = &cobra.Command{
 
 var testCmd = &cobra.Command{
 	Use:   "test",
-	Short: "Run integration tests",
-	Long:  "",
+	Short: "run integration tests",
+	Long:  "run integration tests",
 	Args:  cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return cmdTest(cmd, args)
@@ -327,8 +349,10 @@ func cmdTest(cmd *cobra.Command, args []string) error {
 			func() error { return servertest.DeliverTx(client, []byte{0x00, 0x02}, code.CodeTypeOK, nil) },
 			func() error { return servertest.DeliverTx(client, []byte{0x00, 0x03}, code.CodeTypeOK, nil) },
 			func() error { return servertest.DeliverTx(client, []byte{0x00, 0x00, 0x04}, code.CodeTypeOK, nil) },
-			func() error { return servertest.DeliverTx(client, []byte{0x00, 0x00, 0x06}, code.CodeTypeBadNonce, nil) },
-			func() error { return servertest.Commit(client, []byte{0, 0, 0, 0, 0, 0, 0, 5})},
+			func() error {
+				return servertest.DeliverTx(client, []byte{0x00, 0x00, 0x06}, code.CodeTypeBadNonce, nil)
+			},
+			func() error { return servertest.Commit(client, []byte{0, 0, 0, 0, 0, 0, 0, 5}) },
 		})
 }
 
@@ -496,14 +520,11 @@ func cmdSetOption(cmd *cobra.Command, args []string) error {
 	}
 
 	key, val := args[0], args[1]
-	res, err := client.SetOptionSync(types.RequestSetOption{key, val})
+	_, err := client.SetOptionSync(types.RequestSetOption{key, val})
 	if err != nil {
 		return err
 	}
-	printResponse(cmd, args, response{
-		Code: res.Code,
-		Log:  res.Log,
-	})
+	printResponse(cmd, args, response{Log: "OK (SetOption doesn't return anything.)"}) // NOTE: Nothing to show...
 	return nil
 }
 
@@ -527,6 +548,7 @@ func cmdDeliverTx(cmd *cobra.Command, args []string) error {
 	printResponse(cmd, args, response{
 		Code: res.Code,
 		Data: res.Data,
+		Info: res.Info,
 		Log:  res.Log,
 	})
 	return nil
@@ -537,7 +559,7 @@ func cmdCheckTx(cmd *cobra.Command, args []string) error {
 	if len(args) == 0 {
 		printResponse(cmd, args, response{
 			Code: codeBad,
-			Log:  "want the tx",
+			Info: "want the tx",
 		})
 		return nil
 	}
@@ -552,6 +574,7 @@ func cmdCheckTx(cmd *cobra.Command, args []string) error {
 	printResponse(cmd, args, response{
 		Code: res.Code,
 		Data: res.Data,
+		Info: res.Info,
 		Log:  res.Log,
 	})
 	return nil
@@ -564,9 +587,7 @@ func cmdCommit(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	printResponse(cmd, args, response{
-		Code: res.Code,
 		Data: res.Data,
-		Log:  res.Log,
 	})
 	return nil
 }
@@ -576,7 +597,8 @@ func cmdQuery(cmd *cobra.Command, args []string) error {
 	if len(args) == 0 {
 		printResponse(cmd, args, response{
 			Code: codeBad,
-			Log:  "want the query",
+			Info: "want the query",
+			Log:  "",
 		})
 		return nil
 	}
@@ -596,6 +618,7 @@ func cmdQuery(cmd *cobra.Command, args []string) error {
 	}
 	printResponse(cmd, args, response{
 		Code: resQuery.Code,
+		Info: resQuery.Info,
 		Log:  resQuery.Log,
 		Query: &queryResponse{
 			Key:    resQuery.Key,