+++ types/messages.go
@@ -1,33 +1,70 @@
 package types
 
 import (
+	"bufio"
+	"encoding/binary"
 	"io"
 
 	"github.com/gogo/protobuf/proto"
-	wire "github.com/tendermint/go-wire"
 )
 
-// WriteMessage writes a length-delimited protobuf message.
+const (
+	maxMsgSize = 104857600 // 100MB
+)
+
+// WriteMessage writes a varint length-delimited protobuf message.
 func WriteMessage(msg proto.Message, w io.Writer) error {
 	bz, err := proto.Marshal(msg)
 	if err != nil {
 		return err
 	}
-	var n int
-	wire.WriteByteSlice(bz, w, &n, &err)
-	return err
+	return encodeByteSlice(w, bz)
 }
 
-// ReadMessage reads a length delimited protobuf message.
+// ReadMessage reads a varint length-delimited protobuf message.
 func ReadMessage(r io.Reader, msg proto.Message) error {
-	var n int
-	var err error
-	bz := wire.ReadByteSlice(r, 0, &n, &err) //XXX: no max
+	return readProtoMsg(r, msg, maxMsgSize)
+}
+
+func readProtoMsg(r io.Reader, msg proto.Message, maxSize int) error {
+	// binary.ReadVarint takes an io.ByteReader, eg. a bufio.Reader
+	reader, ok := r.(*bufio.Reader)
+	if !ok {
+		reader = bufio.NewReader(r)
+	}
+	length64, err := binary.ReadVarint(reader)
 	if err != nil {
 		return err
 	}
-	err = proto.Unmarshal(bz, msg)
-	return err
+	length := int(length64)
+	if length < 0 || length > maxSize {
+		return io.ErrShortBuffer
+	}
+	buf := make([]byte, length)
+	if _, err := io.ReadFull(reader, buf); err != nil {
+		return err
+	}
+	return proto.Unmarshal(buf, msg)
+}
+
+//-----------------------------------------------------------------------
+// NOTE: we copied wire.EncodeByteSlice from go-wire rather than keep
+// go-wire as a dep
+
+func encodeByteSlice(w io.Writer, bz []byte) (err error) {
+	err = encodeVarint(w, int64(len(bz)))
+	if err != nil {
+		return
+	}
+	_, err = w.Write(bz)
+	return
+}
+
+func encodeVarint(w io.Writer, i int64) (err error) {
+	var buf [10]byte
+	n := binary.PutVarint(buf[:], i)
+	_, err = w.Write(buf[0:n])
+	return
 }
 
 //----------------------------------------