+++ types/messages_test.go
@@ -8,18 +8,20 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/assert"
+	cmn "github.com/tendermint/tmlibs/common"
 )
 
 func TestMarshalJSON(t *testing.T) {
 	b, err := json.Marshal(&ResponseDeliverTx{})
 	assert.Nil(t, err)
-	assert.True(t, strings.Contains(string(b), "code"))
+	// Do not include empty fields.
+	assert.False(t, strings.Contains(string(b), "code"))
 
 	r1 := ResponseCheckTx{
-		Code: 1,
-		Data: []byte("hello"),
-		Gas:  43,
-		Fee:  12,
+		Code:      1,
+		Data:      []byte("hello"),
+		GasWanted: 43,
+		Fee:       cmn.KI64Pair{[]byte("pho"), 12},
 	}
 	b, err = json.Marshal(&r1)
 	assert.Nil(t, err)
@@ -28,6 +30,26 @@ func TestMarshalJSON(t *testing.T) {
 	err = json.Unmarshal(b, &r2)
 	assert.Nil(t, err)
 	assert.Equal(t, r1, r2)
+}
+
+func TestWriteReadMessageSimple(t *testing.T) {
+	cases := []proto.Message{
+		&RequestEcho{
+			Message: "Hello",
+		},
+	}
+
+	for _, c := range cases {
+		buf := new(bytes.Buffer)
+		err := WriteMessage(c, buf)
+		assert.Nil(t, err)
+
+		msg := new(RequestEcho)
+		err = ReadMessage(buf, msg)
+		assert.Nil(t, err)
+
+		assert.Equal(t, c, msg)
+	}
 }
 
 func TestWriteReadMessage(t *testing.T) {
@@ -44,6 +66,34 @@ func TestWriteReadMessage(t *testing.T) {
 		assert.Nil(t, err)
 
 		msg := new(Header)
+		err = ReadMessage(buf, msg)
+		assert.Nil(t, err)
+
+		assert.Equal(t, c, msg)
+	}
+}
+
+func TestWriteReadMessage2(t *testing.T) {
+	phrase := "hello-world"
+	cases := []proto.Message{
+		&ResponseCheckTx{
+			Data:      []byte(phrase),
+			Log:       phrase,
+			GasWanted: 10,
+			Tags: []cmn.KVPair{
+				cmn.KVPair{[]byte("abc"), []byte("def")},
+			},
+			// Fee: cmn.KI64Pair{
+		},
+		// TODO: add the rest
+	}
+
+	for _, c := range cases {
+		buf := new(bytes.Buffer)
+		err := WriteMessage(c, buf)
+		assert.Nil(t, err)
+
+		msg := new(ResponseCheckTx)
 		err = ReadMessage(buf, msg)
 		assert.Nil(t, err)