+++ types/util.go
@@ -3,15 +3,12 @@ package types
 import (
 	"bytes"
 	"encoding/json"
-
-	"github.com/tendermint/go-wire/data"
-	cmn "github.com/tendermint/tmlibs/common"
 )
 
 //------------------------------------------------------------------------------
 
 // Validators is a list of validators that implements the Sort interface
-type Validators []*Validator
+type Validators []Validator
 
 func (v Validators) Len() int {
 	return len(v)
@@ -31,36 +28,16 @@ func (v Validators) Swap(i, j int) {
 func ValidatorsString(vs Validators) string {
 	s := make([]validatorPretty, len(vs))
 	for i, v := range vs {
-		s[i] = validatorPretty{v.PubKey, v.Power}
+		s[i] = validatorPretty(v)
 	}
 	b, err := json.Marshal(s)
 	if err != nil {
-		cmn.PanicSanity(err.Error())
+		panic(err.Error())
 	}
 	return string(b)
 }
 
 type validatorPretty struct {
-	PubKey data.Bytes `json:"pub_key"`
-	Power  int64      `json:"power"`
-}
-
-//------------------------------------------------------------------------------
-
-// KVPairInt is a helper method to build KV pair with an integer value.
-func KVPairInt(key string, val int64) *KVPair {
-	return &KVPair{
-		Key:       key,
-		ValueInt:  val,
-		ValueType: KVPair_INT,
-	}
-}
-
-// KVPairString is a helper method to build KV pair with a string value.
-func KVPairString(key, val string) *KVPair {
-	return &KVPair{
-		Key:         key,
-		ValueString: val,
-		ValueType:   KVPair_STRING,
-	}
+	PubKey []byte `json:"pub_key"`
+	Power  int64  `json:"power"`
 }